+++ aws-java-sdk-code-generator/src/main/java/com/amazonaws/codegen/customization/processors/ShapeModifiersProcessor.java
@@ -27,9 +27,12 @@ import com.amazonaws.codegen.model.service.Member;
 import com.amazonaws.codegen.model.service.ServiceModel;
 import com.amazonaws.codegen.model.service.Shape;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Set;
 
 /**
  * This processor handles all the modification on the shape members in the
@@ -190,6 +193,15 @@ final class ShapeModifiersProcessor implements CodegenCustomizationProcessor {
                 }
                 shape.getMembers().putAll(injects);
             }
+        }
+
+        if (modifier.getInjectEnumValues() != null) {
+            if (shape.getEnumValues() == null) {
+                throw new IllegalStateException("Attempted to add enum members to a non-enum shape!");
+            }
+            Set<String> extendValues = new HashSet<>(shape.getEnumValues());
+            extendValues.addAll(modifier.getInjectEnumValues());
+            shape.setEnumValues(new ArrayList<>(extendValues));
         }
     }